requirements.txt
Django==4.1.4
numpy==1.25.1
Pillow==9.2.0
Pillow==10.0.0
tensorflow==2.13.0
